Remove the separate file for authorize method.
