+++ front-end/src/lib/auth.ts
@@ -1,8 +1,20 @@
-import { authorize } from "@/lib/signInApi";
 import { User } from "@/types/user";
 import { JWT } from "@auth/core/jwt";
 import Credentials from "@auth/core/providers/credentials";
+import axios, { AxiosError } from "axios";
 import NextAuth, { Session } from "next-auth";
+
+export interface ApiUser {
+  UserID: number;
+  UserFirstName: string;
+  UserLastName: string;
+  email: string;
+  DateHired: Date;
+  UserRoleID: number;
+  RoleName: string;
+  created_at: string | null;
+  updated_at: string | null;
+}
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
   providers: [
@@ -15,7 +27,34 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
           type: "password",
         },
       },
-      authorize,
+      authorize: async (credentials) => {
+        try {
+          const response = await axios.post<ApiUser>(
+            "/auth/login",
+            credentials,
+          );
+          const apiUser = response.data;
+
+          return {
+            id: apiUser.UserID.toString(),
+            firstName: apiUser.UserFirstName,
+            lastName: apiUser.UserLastName,
+            email: apiUser.email,
+            dateHired: new Date(apiUser.DateHired),
+            roleId: apiUser.UserRoleID,
+            roleName: apiUser.RoleName,
+          };
+        } catch (error) {
+          if (
+            !(error instanceof AxiosError) ||
+            error.response?.status !== 401
+          ) {
+            console.error(error);
+          }
+
+          return null;
+        }
+      },
     }),
   ],
   callbacks: {
